backend/requirements.txt
fastapi
uvicorn
PyJWT
passlib
python-multipart
python-dotenv
psycopg2-binary
sqlalchemy

